backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.0
pinecone-client==2.2.4
pypdf==3.17.1
python-docx==0.8.11
tiktoken==0.5.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
cors-headers==4.3.1
